pyproject.toml
[project]
name = "eratosthenes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "datasets>=3.6.0",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-sqlalchemy>=3.1.1",
    "jsonlines>=4.0.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "nltk>=3.9.1",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "pyvis>=0.3.2",
    "rouge-score>=0.1.2",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "trl>=0.17.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
